Extract letter links.
